ipacm: Changes to handle calculation overflow
Added changes to handle IPACM file size calculation overflow while
calculating the max file size.

+++ ipacm/src/IPACM_Log.cpp
@@ -63,7 +63,7 @@
 char* dump_file = 0;
 void* mmap_addr = NULL;
 void* write_addr = 0;
-uint32_t max_filesize = 0;
+int64_t max_filesize = 0;
 int log_init_done = 0;
 pthread_mutex_t file_lock;
 
@@ -238,7 +238,7 @@ int log_init() {
 	}
 	if(is_exist == false)
 	{
-		trunc_ret = ftruncate(log_fd, max_filesize);
+		trunc_ret = ftruncate(log_fd, (off_t)max_filesize);
 		if(0 > trunc_ret)
 		{
 			perror("Ftruncate failed\n");
